move InternalKeyComparator to public header so it can be inlined
