+++ port/win/port_win.h
@@ -108,7 +108,7 @@ const size_t kMaxSizet = UINT_MAX;
 
 #endif //_MSC_VER
 
-const bool kLittleEndian = true;
+extern const bool kLittleEndian;
 
 class CondVar;
