+++ util/perf_context_imp.h
@@ -44,10 +44,6 @@ namespace rocksdb {
 #define PERF_TIMER_MEASURE(metric)        \
   perf_step_timer_ ## metric.Measure();
 
-// Increase metric value
-#define PERF_COUNTER_ADD(metric, value)     \
-  perf_context.metric += value;
-
 #endif
 
-}
+}  // namespace rocksdb
